added timer context to add timer logic
